+++ examples/detached_mode.py
@@ -32,7 +32,7 @@ import sys
 import time
 
 import parsl
-from parsl.app.python import python_app
+from parsl import python_app
 from parsl.config import Config
 from parsl.executors import HighThroughputExecutor