+++ examples/serverless_mode.py
@@ -38,7 +38,7 @@ import os
 import sys
 
 import parsl
-from parsl.app.python import python_app
+from parsl import python_app
 from parsl.config import Config
 from parsl.executors import HighThroughputExecutor