+++ examples/spot_fleet_example.py
@@ -29,8 +29,8 @@ import os
 import sys
 
 import parsl
+from parsl import python_app
 from parsl.addresses import address_by_route
-from parsl.app.app import python_app
 from parsl.config import Config
 from parsl.executors import HighThroughputExecutor